+++ src/Moment/MomentGenerator.h
@@ -14,7 +14,7 @@
 #include <chrono>
 #include <thread>
 
-#include "../InterfaceConstants.h"
+#include "../Constants.h"
 #include "ImageMoments.h"
 
 using MomentProgressCallback = std::function<void(float)>;
