+++ src/Session.cc
@@ -28,12 +28,12 @@
 #include <carta-protobuf/raster_tile.pb.h>
 
 #include "Carta.h"
+#include "Constants.h"
 #include "DataStream/Compression.h"
 #include "EventHeader.h"
 #include "FileList/FileExtInfoLoader.h"
 #include "FileList/FileInfoLoader.h"
 #include "FileList/FitsHduList.h"
-#include "InterfaceConstants.h"
 #include "OnMessageTask.h"
 #include "SpectralLine/SpectralLineCrawler.h"
 #include "Timer/Timer.h"
